+++ pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 testpaths = tests
-;addopts = -n logical
+addopts = -n logical
 ; Use below option to temporarily disable pytest-xdist. See https://yossarian.net/til/post/disabling-pytest-xdist-without-removing-it/
- addopts = -n 0
+;  addopts = -n 0
 
 ;addopts = -q --tb=line
