+++ tests/complex_model_suite/test_umep_processing.py
@@ -2,9 +2,10 @@ import os
 import pytest
 from pytest_check import equal # https://github.com/okken/pytest-check
 
-from src.constants import DATA_DIR, FOLDER_NAME_PRIMARY_DATA, FOLDER_NAME_PRIMARY_RASTER_FILES
+from src.constants import DATA_DIR
 from src.workers.city_data import CityData
 from src.workers.worker_tools import remove_folder, create_folder
+from tests.conftest import RUN_COMPLEX_MODEL_SUITE
 from tests.testing_tools import run_main, SAMPLE_CITIES_SOURCE_DIR, file_count_in_vrt_directory, \
     does_file_signature_match, get_geotiff_signature
 
@@ -13,11 +14,8 @@ CLEANUP_RESULTS=False
 SCRATCH_TARGET_DIR = os.path.join(DATA_DIR, 'scratch_target')
 create_folder(SCRATCH_TARGET_DIR)
 
-# TODO Add test fpr era5
 
-RUN_CORE_TESTS_ONLY = False
-
-
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_tropical_full_cif_city():
     source_city_folder_name = 'BRA_Rio_de_Janeiro_full_cif'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -42,7 +40,7 @@ def test_tropical_full_cif_city():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_tiled_cif_city():
     source_city_folder_name = 'NLD_Amsterdam'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -67,7 +65,7 @@ def test_tiled_cif_city():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_tiled_buffered_cif_city_umep():
     source_city_folder_name = 'NLD_Amsterdam_buftile_cif_umep'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -92,7 +90,7 @@ def test_tiled_buffered_cif_city_umep():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_tiled_custom_city():
     source_city_folder_name = 'NLD_Amsterdam_custom_tiled'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -117,7 +115,7 @@ def test_tiled_custom_city():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_portland_swath_cif_umep():
     source_city_folder_name = 'USA_Portland_swath_cif_umep'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -141,7 +139,8 @@ def test_portland_swath_cif_umep():
         if CLEANUP_RESULTS:
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_USA_WashingtonDC_cif_umep():
     source_city_folder_name = 'USA_WashingtonDC_cif_umep'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -165,7 +164,8 @@ def test_USA_WashingtonDC_cif_umep():
         if CLEANUP_RESULTS:
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_download_only_cif_city():
     source_city_folder_name = 'ZAF_Capetown_cif_download_only'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -185,6 +185,7 @@ def test_download_only_cif_city():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_ZAF_Capetown_cif_local_era5_umep():
     source_city_folder_name = 'ZAF_Capetown_cif_local_era5_umep'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -209,7 +210,7 @@ def test_ZAF_Capetown_cif_local_era5_umep():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_mixed_custom_city():
     source_city_folder_name = 'ZAF_Capetown_small_mixed_cif'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -234,6 +235,7 @@ def test_mixed_custom_city():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_custom_city():
     source_city_folder_name = 'ZAF_Capetown_small_tile'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -258,7 +260,7 @@ def test_custom_city():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_custom_city_with_full_intermediates():
     source_city_folder_name = 'ZAF_Capetown_with_full_intermediates'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -283,7 +285,7 @@ def test_custom_city_with_full_intermediates():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
-@pytest.mark.skipif(RUN_CORE_TESTS_ONLY == True, reason='Skipping since RUN_CORE_TESTS_ONLY set to True')
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_custom_city_with_mixed_intermediates():
     source_city_folder_name = 'ZAF_Capetown_with_mixed_intermediates'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -308,6 +310,7 @@ def test_custom_city_with_mixed_intermediates():
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
 
+@pytest.mark.skipif(RUN_COMPLEX_MODEL_SUITE is False, reason=f"Skipping since RUN_FULL_TEST_SUITE set to {RUN_COMPLEX_MODEL_SUITE}")
 def test_USA_Philadelphia_downtown_cif_umep():
     source_city_folder_name = r'USA_Philadelphia_downtown_cif_umep'
     non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
@@ -331,28 +334,3 @@ def test_USA_Philadelphia_downtown_cif_umep():
         if CLEANUP_RESULTS:
             remove_folder(non_tiled_city_data.target_city_parent_path)
 
-
-def test_ZAF_Capetown_cif_umep():
-    source_city_folder_name = r'ZAF_Capetown_cif_umep'
-    non_tiled_city_data = CityData(source_city_folder_name, None, SAMPLE_CITIES_SOURCE_DIR, SCRATCH_TARGET_DIR)
-    try:
-        remove_folder(non_tiled_city_data.target_city_parent_path)
-        return_code = run_main(SCRATCH_TARGET_DIR, source_city_folder_name, 'run_pipeline')
-
-        mrt_file_name = 'Tmrt_2023_1_1200D.tif'
-        target_met_file = os.path.join(non_tiled_city_data.target_tcm_results_path, 'met_era5_hottest_days', 'tile_001',
-                                       mrt_file_name)
-        expected_signature = get_geotiff_signature(target_met_file)
-
-        # expected_signature =
-        mrt_signatures_are_equal = does_file_signature_match(expected_signature, target_met_file)
-        assert mrt_signatures_are_equal
-
-        vrt_count = file_count_in_vrt_directory(non_tiled_city_data)
-
-        assert return_code == 0
-        assert vrt_count == 17
-
-    finally:
-        if CLEANUP_RESULTS:
-            remove_folder(non_tiled_city_data.target_city_parent_path)
