+++ tests/conftest.py
@@ -1,0 +1,1 @@
+RUN_COMPLEX_MODEL_SUITE = False
